+++ modules/runtime/src/misc/log/log_worker.hpp
@@ -1,9 +1,10 @@
 #pragma once
+#include "async/async_service.h"
 
 namespace skr {
 namespace log {
 
-struct LogWorker
+struct LogWorker : public ServiceThread
 {
     void run()
     {
